# swagger/portal-api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: wicked.haufe.io Portal API
  description: The API which the API Portal uses to store and retrieve information
 
# Under securityDefinitions we declare which security schemes can be used.
# These definitions do not affect to the operations, but can be applied
# using the 'security' key at the global level or the operation level.
# In this sample, we'll see how to apply the security definition globally,
# and override it at the operation level.
#
# Note that for Access Code Flow, you must provide the Token URL.
securityDefinitions:
  # For personalized access to data
  accessCode:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/auth/local/portal-api/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/auth/local/portal-api/oauth2/authorize
    flow: accessCode
    scopes:
      read_plans: Read plans
      read_groups: Read user group definitions
      read_content: Read custom/markdown/HTML content
      health: Read out health information
      read_templates: Read templates
      read_auth_servers: Read authorization servers
      read_apis: Read API definitions
      read_users: Read user information
      write_users: Write user information
      read_applications: Read application data
      write_applications: Write application data
      read_subscriptions: Read subscriptions
      write_subscriptions: Write/edit/delete subscriptions
      read_approvals: Read pending approvals
      read_verifications: Read email and password verifications
      write_verifications: Create or delete verifications
      read_registrations: Read user registrations
      write_registrations: Write user registrations
      read_grants: Read user API grants
      write_grants: Write user API grants
      webhooks: Access and change webhooks
  
  implicit:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/portal-api/v1/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/portal-auth/api/portal-api
    flow: accessCode
    scopes:
      read_plans: Read plans
...
produces:
  - application/hal+json
host: 'localhost:3001'
basePath: /portal-api/v1
...
paths:
...
  '/webhooks/listeners':
    get:
      security:
      - accessCode: ['webhooks']
      - implicit: ['webhooks']
      - password: ['webhooks']
      description: Read all webhook listeners (admin)
      tags:
        - webhooks
      responses:
        '200':
          description: Returns a list of registered webhook listeners
          schema:
            type: array
            items:
              $ref: '#/definitions/listenerType'
        '403':
          description: Not allowed
  '/webhooks/listeners/{listenerId}':
    put:
      security:
      - accessCode: ['webhooks']
      - implicit: ['webhooks']
      - password: ['webhooks']
      description: Update/create a webhook listener (admin)
      tags:
        - webhooks
      parameters:
        - name: listenerId
          in: path
          type: string
          required: true
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/listenerType'
      responses:
        '200':
          description: Success (update or create)
        '400':
          description: Bad request
    delete:
      security:
      - accessCode: ['webhooks']
      - implicit: ['webhooks']
      - password: ['webhooks']
      description: Deletes a webhook listener (admin)
      tags:
        - webhooks
      parameters:
        - name: listenerId
          in: path
          type: string
          required: true
      responses:
        '204':
          description: 'Success, no content'
        '403':
          description: Not allowed
        '404':
          description: Not found
  '/webhooks/events/{listenerId}':
    get:
      security:
      - accessCode: ['webhooks']
      - implicit: ['webhooks']
      - password: ['webhooks']
      tags:
        - webhooks
      description: Gets the list of pending events (admin)
      parameters:
        - name: listenerId
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              title: WebhookEvent
              type: object
              required:
                - id
                - action
                - entity
              properties:
                id:
                  type: string
                action:
                  type: string
                entity:
                  type: string
                href:
                  type: string
                data:
                  description: Any kind of data
                  type: object
    delete:
      security:
      - accessCode: ['webhooks']
      - implicit: ['webhooks']
      - password: ['webhooks']
      tags:
        - webhooks
      description: Flushes the event queue for this listener (admin)
      parameters:
        - name: listenerId
          in: path
          type: string
          required: true
      responses:
        '204':
          description: 'Successful flush, no content'
        '403':
          description: Not allowed
  '/webhooks/events/{listenerId}/{eventId}':
    delete:
      security:
      - accessCode: ['webhooks']
      - implicit: ['webhooks']
      - password: ['webhooks']
      tags:
        - webhooks
      description: Acknowledge/delete an event (admin)
      parameters:
        - name: listenerId
          in: path
          type: string
          required: true
        - name: eventId
          in: path
          type: string
          required: true
      responses:
        '204':
          description: Successfully deleted
        '403':
          description: Not allowed
...
definitions:
...
  listenerType:
    type: object
    properties:
      id:
        type: string
      url:
        type: string
...